Made design changes to make different pages more coherent.
Notice that the code behind connect directly to the database when doing the "on get":s.

Didn't change it since it's not a big priority at the moment

+++ Pages/Customers/Details.cshtml.cs
@@ -12,7 +12,8 @@ namespace Retro_grupp_g.Pages.Customers
         public Customer Customer {  get; set; } = default!;
         public async Task<IActionResult> OnGetAsync(int id)
         {
-            var customer = await _repo.GetDetailsAsync((ushort)id);
+            var customer = await _repo.GetDetailsAsync(id);
+
             if (customer == null) return NotFound();
 
             Customer = customer;
